% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CompareAllTrees}
\alias{CompareAllTrees}
\title{All distances between each pair of trees}
\usage{
CompareAllTrees(trees, exact = FALSE)
}
\arguments{
\item{trees}{List of bifurcating trees of class \code{phylo}.}

\item{exact}{Logical specifying whether to calculate exact rearrangement
distances.}
}
\description{
All distances between each pair of trees
}
\examples{
  trees <- lapply(rep(8, 5), ape::rtree, br = NULL)
  CompareAllTrees(trees)

}
\seealso{
Other pairwise tree distances: \code{\link{AllDists}},
  \code{\link{PairwiseDistances}}
}
\author{
Martin R. Smith

Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{pairwise tree distances}
