% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distanceDistribution25}
\alias{distanceDistribution25}
\title{Distances between random pairs of trees}
\format{An object of class \code{matrix} with 9 rows and 10000 columns.}
\usage{
distanceDistribution25
}
\description{
Two-dimensional matrices listing the normalized distances between random
pairs of 11- and 25-tip trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br=NULL)}.
}
\details{
\code{pectinateDistances11} reports distances between random trees and a pectinate
11-tip tree.

Rows are named with an abbreviation of the tree comparison metric.
Variation of information measures are normalized against the maximum
possible variation of information for trees of the corresponding topologies.
The quartet distance and Nye \emph{et al.} measures are normalized against their
maximum possible values.  The remaining measures are unnormalized.
\itemize{
\item \code{vpi}: Variation of Phylogenetic Information (Smith, forthcoming)
\item \code{vmsi}: Variation of Matching Split Information (Smith, forthcoming)
\item \code{vci}: Variation of Clustering Information (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012), unnormalized
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1985), unnormalized
\item \code{path}: Path distance (Steel & Penny 1993), unnormalized
\item \code{spr}: SPR distance, unnormalized
}

Each column lists the calculated distances between each pair of trees.

The pairs of 25-tip trees are saved as data object \code{\link{randomTreePairs25}}.
}
\references{
\insertRef{Bogdanowicz2012}{TreeDist}
\insertRef{Nye2006}{TreeDist}
\insertRef{Robinson1981}{TreeDist}
\insertRef{SmithTern}{TreeSearch}
\insertRef{SmithDist}{TreeDist}
\insertRef{Steel1993}{TreeDist}
}
\keyword{datasets}
