% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distanceDistributions}
\alias{distanceDistributions}
\alias{distanceDistribution50}
\alias{distanceDistribution25}
\alias{pectinateDistances11}
\title{Distances between random pairs of trees}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 21 rows and 10000 columns.

An object of class \code{matrix} (inherits from \code{array}) with 21 rows and 10000 columns.

An object of class \code{matrix} (inherits from \code{array}) with 22 rows and 100000 columns.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
distanceDistribution50

distanceDistribution25

pectinateDistances11
}
\description{
Two-dimensional matrices listing the normalized distances between random
pairs of bifurcating trees with 11, 25 and 50 leaves drawn from the
uniform distribution using \code{ape::rtree(nTip, br=NULL)}.
}
\details{
\code{pectinateDistances11} reports distances between a pectinate 11-leaf tree
and 100 000 random bifurcating trees.

Rows are named with an abbreviation of the tree comparison metric.
Information-based measures are normalized against the splitwise information
content/entropy for binary trees of the corresponding topologies.<!--#TODO VERIFY-->


The quartet distance, Nye \emph{et al.} and Jaccard-Robinson-Foulds measures
are normalized against their maximum possible values.
The remaining measures are unnormalized.
\itemize{
\item \code{dpi}: Different Phylogenetic Information (Smith, forthcoming)
\item \code{msid}: Matching Split Information Distance (Smith, forthcoming)
\item \code{cid}: Clustering Information Distance (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{ja2}, \code{ja4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4
\item \code{jna2}, \code{jna4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
with non-arboreal matchings permitted
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012), unnormalized
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009)
\item \code{masti}: Information content of Maximum Agreement Subtree
\item \code{nni_l}, \code{nni_t}, \code{nni_u}: Lower bound, tight upper bound, and upper bound
for  nearest-neighbour interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate SPR distance, unnormalized
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(TBR) distance
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1985), unnormalized
\item \code{rfi}: Robinson-Foulds distance, splits weighted by phylogenetic
information content (Smith, forthcoming)
\item \code{path}: Path distance (Steel & Penny 1993), unnormalized
}

Each column lists the calculated distances between each pair of trees.

The pairs of 25-leaf trees are saved as data object \code{\link{randomTreePairs25}}.
}
\references{
\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}
}
\keyword{datasets}
