% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.TDDTable}
\alias{.TDDTable}
\title{Tabulate method outputs}
\usage{
.TDDTable(Table, dat, ...)
}
\arguments{
\item{Table}{\code{DT::DataTable()} or an equivalent tabulation function}

\item{dat}{Matrix of data to be tabulated, with row names corresponding to
the methods employed.}
}
\value{
\code{.TDDTable()} produces a table plotted using \code{Table()}.
}
\description{
Helper function to tabulate sortable data in package vignettes.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
