% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{KCDiameter}
\alias{KCDiameter}
\title{Approximate diameter of the Kendall–Collijn metric}
\usage{
KCDiameter(nTip)
}
\arguments{
\item{nTip}{Integer specifying number of leaves.}
}
\value{
\code{KCDiameter()} returns a single numeric.
}
\description{
\code{KCDiameter()} calculates the value of the Kendall & Colijn's (2016)
metric distance between two pectinate trees with \code{nTip} leaves ordered in
the opposite direction, which I suggest (without any attempt at a proof) may
be a useful proxy for the diameter (i.e. maximum value) of the K–C
metric.
}
\examples{
KCMax(4)
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
