% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{KCMax}
\alias{KCMax}
\title{Approximate diameter of the Kendall—Collijn metric}
\usage{
KCMax(nTip)
}
\arguments{
\item{nTip}{Integer specifying number of leaves.}
}
\value{
\code{KXMax()} returns a single numeric.
}
\description{
\code{KCMax()} calculates the value of the Kendall-Colijn (2016) metric distance
between two pectinate trees with \code{nTip} leaves ordered in the opposite
direction, which I have suggested (without any attempt at a proof) may
be a useful proxy for the diameter (i.e. maximum value) of the KC metric.
}
\examples{
KCMax(4)
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
