% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linTests}
\alias{linTests}
\alias{linTestOneResults}
\alias{linTestTwoResults}
\alias{linTestSPRResults}
\title{Evaluating tree distance metrics by cluster recovery}
\format{An object of class \code{array} of dimension 5 x 18 x 5.}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
linTestOneResults

linTestTwoResults

linTestSPRResults
}
\description{
An effective measure of tree distance will recover clusters of similar
trees.  These datasets contain the results of tests modelled on those
in Lin \emph{et al.} (2012).
}
\details{
I used three approaches to generate clusters of similar trees, and tested
each metric in its ability to recover these clusters (Lin et al., 2012).

For the first test, I generated 500 datasets of 100 binary trees with
\emph{n} = 40 leaves.
Each set of trees was created by randomly selecting two \emph{k}-leaf
'skeleton' trees, where \emph{k} ranges from 0.3 \emph{n} to 0.9 \emph{n}.
From each skeleton, 50 trees were generated by adding each of the remaining
\emph{n} - \emph{k} leaves in turn at a uniformly selected point on the tree.

For the second and third test, each dataset was constructed by selecting at
random two binary 40-leaf trees.
From each starting tree, I generated 50 binary trees by conducting \emph{k}
leaf-label interchange (LLI) operations (test two) or \emph{k} subtree prune and
regraft (SPR) operations (test three) on the starting tree.
An LLI operation swaps the positions of two randomly selected leaves,
without affecting tree shape; an SPR operation moves a subtree to a new
location within the tree.

For each dataset, I calculated the distance between each pair of trees.
Trees where then partitioned into clusters using five methods,
using the \pkg{stats} and \pkg{cluster}: spectral clustering,
partitioning around medioids, and hierarchical clustering using complete,
single and average linkage.
I define the success rate of each distance measure as the proportion of
datasets in which every tree generated from the same skeleton was placed
in the same cluster.
}
\references{
\insertRef{Lin2012}{TreeDistData}
}
\keyword{datasets}
