% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{randomTreeDistances}
\alias{randomTreeDistances}
\title{Mean distances between random pairs of trees}
\format{
An object of class \code{array} of dimension 20 x 13 x 197.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
randomTreeDistances
}
\description{
A three-dimensional array listing the normalized distances between 1 000
random pairs of trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br = NULL)}.
}
\details{
Normalization is against the maximum possible value obtainable on a pair
of trees of the shapes given, with the exception of the SPR distance,
which is normalized against the upper bound (\code{spr}) and lower bound (\code{sprLB})
of its diameter on a tree with \emph{n} leaves.  The path and matching split
distances are not normalized.

Rows are named with abbreviations of the tree comparison metrics tested
(see 'Methods tested' below).

Columns list the summary statistics of calculated tree distances: the
minimum (\code{min}),
1\%, 5\%, 10\%, 25\%, 50\% (i.e. median), 75\%, 90\%, 95\%, 99\% percentiles,
maximum (\code{max}), mean (\code{mean}) and standard deviation (\code{sd}).

The third dimension lists the number of leaves in the trees compared.

For analysis of this data, see the
\href{https://ms609.github.io/TreeDistData/articles/09-expected-similarity.html}{accompanying vignette}.
}
\section{Methods tested}{
\itemize{
\item \code{pid}: Phylogenetic Information Distance (Smith 2020)
\item \code{msid}: Matching Split Information Distance (Smith 2020)
\item \code{cid}: Clustering Information Distance (Smith 2020)
\item \code{qd}: Quartet divergence (Smith 2019)
\item \code{nye}: Nye \emph{et al.} tree distance (Nye \emph{et al.} 2006)
\item \code{jnc2}, \code{jnc4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings prohibited ('no-conflict')
\item \code{joc2}, \code{jco4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings permitted ('conflict-ok')
\item \code{ms}: Matching Split Distance (Bogdanowicz & Giaro 2012)
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009)
\item \code{masti}: Information content of Maximum Agreement Subtree
\item \code{nni_l}, \code{nni_u}: Lower
and upper bounds
for nearest-neighbour interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate SPR distance
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(TBR) distance, calculated using
\href{https://ms609.github.io/TBRDist/}{\pkg{TBRDist}}
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1981)
\item \code{icrf}: Information-corrected Robinson-Foulds distance (Smith 2020)
\item \code{path}: Path distance (Steel & Penny 1993)
}
}

\references{
\itemize{
\item \insertRef{Allen2001}{TreeDist}
\item \insertRef{Bocker2013}{TreeDist}
}

\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Kendall2016}{TreeDistData}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}
}
\keyword{datasets}
