% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{randomTreePairs}
\alias{randomTreePairs}
\alias{randomTreePairs25}
\alias{randomTreePairs50}
\title{Pairs of random trees}
\format{
An object of class \code{list} of length 10000.

An object of class \code{list} of length 10000.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
randomTreePairs25

randomTreePairs50
}
\description{
Lists of 10 000 pairs of binary trees drawn from the uniform
distribution using \code{ape::rtree(nTip, br = NULL)}.
}
\seealso{
The distances between these pairs of trees are recorded in
the data objects  \code{\link{distanceDistribution25}} and  \code{\link{distanceDistribution50}}.
}
\keyword{datasets}
