% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomTreePairs25}
\alias{randomTreePairs25}
\title{Pairs of random 25-tip trees}
\format{An object of class \code{list} of length 10000.}
\usage{
randomTreePairs25
}
\description{
A list of 10 000 25-tip trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br=NULL)}.
}
\details{
The distances between these pairs of trees are recorded in
the data object \code{\link{distanceDistribution25}}.
}
\keyword{datasets}
