% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sprDistances}
\alias{sprDistances}
\title{SPR walk distances}
\format{An object of class \code{list} of length 21.}
\usage{
sprDistances
}
\description{
<<#TODO Describe in detail.>>
}
\keyword{datasets}
