% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{treeBalance}
\alias{treeBalance}
\alias{balance25}
\alias{balance50}
\title{Correlation between tree distances and tree balance}
\format{
An object of class \code{numeric} of length 25.

An object of class \code{numeric} of length 25.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
balance25

balance50
}
\description{
The balance of each tree in the lists \code{\link{distanceDistribution25}} and
\code{\link{distanceDistribution50}} was quantified using the total cophenetic
index (Mir \emph{et al.} 2013).
The difference in balance for each pair of trees was then correlated with
the distance between those trees, and the r² value recorded.
}
\references{
\insertRef{Mir2013}{TreeTools}
}
\keyword{datasets}
