% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TreeDistCol}
\alias{TreeDistCol}
\title{Select colour from palette}
\usage{
TreeDistCol(method, opacity = "FF")
}
\arguments{
\item{method}{Character specifying acronym for method; one of \code{tdMethods}.}

\item{opacity}{Character specifying fex code for opacity; \code{"FF"} = opaque.}
}
\description{
Select colour from palette
}
