% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeDistData-package.R
\docType{package}
\name{TreeDistData-package}
\alias{TreeDistData}
\alias{TreeDistData-package}
\title{'TreeDistData' is an R package that contains the functions and scripts used to
compare the performance of tree distance metrics in Smith (2020).}
\description{
The "\href{https://ms609.github.io/TreeDistData/articles/}{Articles}" link in the
navigation bar lists vignettes detailing the output of analyses corresponding
to the stated sections of Smith (2020).  Each of these pages includes a
link to the R source code, which can be used to reproduce the analytical
results.
}
\details{
The "\href{https://ms609.github.io/TreeDistData/reference/}{Reference}" link lists
convenience functions, which facilitate calls to "TreeDist", and data objects
that store the results of lengthy analyses.  The code used to generate
these data objects is available in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
\subsection{References}{

Smith, M.R. (2020). \href{https://dx.doi.org/10.1093/bioinformatics/btaa614}{Information theoretic Generalized Robinson-Foulds metrics for comparing phylogenetic trees}. \emph{Bioinformatics},
online ahead of print.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeDistData/}
  \item \url{https://github.com/ms609/TreeDistData/}
  \item Report bugs at \url{https://github.com/ms609/TreeDistData/issues/}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [copyright holder]

}
\keyword{internal}
