% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TreeDistMethods}
\alias{TreeDistMethods}
\alias{tdAbbrevs}
\alias{tdMethods}
\alias{tdCol}
\alias{TDFunctions}
\alias{TDPair}
\title{Method parameters}
\format{An object of class \code{expression} of length 26.}
\usage{
tdAbbrevs

tdMethods

tdCol

TDFunctions

TDPair
}
\description{
<<#TODO Describe in detail.>>
}
\keyword{datasets}
