% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AllDists}
\alias{AllDists}
\title{All distances between a pair of trees}
\usage{
AllDists(tr1, tr2, verbose = FALSE)
}
\arguments{
\item{tr1, tr2}{Phylogenetic trees of class \code{phylo}.}

\item{verbose}{Logical specifying whether to notify user of progress.}
}
\description{
All distances between a pair of trees
}
\seealso{
Other pairwise tree distances: \code{\link{CompareAllTrees}},
  \code{\link{PairwiseDistances}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{pairwise tree distances}
