% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AllDists}
\alias{AllDists}
\title{All distances between a pair of trees}
\usage{
AllDists(tr1, tr2)
}
\arguments{
\item{tr1, tr2}{Phylogenetic trees of class \code{phylo}.}
}
\description{
All distances between a pair of trees
}
\author{
Martin R. Smith
}
\keyword{internal}
