% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullseye}
\alias{bullseye}
\alias{bullseyeTrees}
\alias{bullMoDiInferred}
\alias{bullMoDiScores}
\alias{bullseyeMorphInferred}
\alias{bullseyeMorphScores}
\alias{bullseyeDistances}
\title{Bullseye test results}
\format{An object of class \code{list} of length 4.}
\usage{
bullseyeTrees

bullMoDiInferred

bullMoDiScores

bullseyeMorphInferred

bullseyeMorphScores

bullseyeDistances
}
\description{
Implementation and results of a 'Bullseye' test, after that proposed by
Kuhner and Yamato (2015).
}
\details{
\code{bullseyeTrees} is a list with four elements, named \code{5 leaves}, \code{10 leaves},
\code{20 leaves} and \code{50 leaves}.  Each element contains 1 000 trees with \emph{n}
leaves, randomly sampled from the uniform distribution using \code{ape::rtree}.

\code{bullXXInferred} is a list with four elements, named as in \code{bullseyeTrees}.
Each element contains 1 000 subelements. Each subelement is a list of
ten trees, which have been inferred from progressively more degraded datasets,
originally simulated from the corresponding tree in \code{bullseyeTrees}.

\code{bullXXScores} is a list with four elements, named as in \code{bullseyeTrees}.
Each element contains a three dimensional array, in which the first dimension
corresponds to the progressive degrees of degradation, labelled according to
the number of characters present or the percentage of tokens switched;
the second dimension is named with an abbreviation of the tree similarity /
distance metric used to score the trees, and the third dimension contains
1 000 entries corresponding to the trees in \code{bullseyeTrees}.
Each cell contains the distance between the inferred tree and the generative
tree under the stated tree distance metric.

The \code{bullseyeMorph} prefix refers to the 'subsampling' experiment
described by Smith (forthcoming); the \code{bullMoDi} prefix refers to the
'miscoding' experiment.

\code{bullseyeDistances} contains four elements, each tabulating the distance
between each pair of \emph{n}-leaf trees in \code{bullseyeTrees}.  For details, see
\code{\link{sevenTipDistances}}.
}
\references{
\itemize{
\item \insertRef{Kuhner2015}{TreeDistData}
\item \insertRef{SmithDist}{TreeDist}
}
}
\keyword{datasets}
