% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CompareAllTrees}
\alias{CompareAllTrees}
\title{All distances between each pair of trees}
\usage{
CompareAllTrees(trees)
}
\arguments{
\item{trees}{List of bifurcating trees of class \code{phylo}.}
}
\description{
All distances between each pair of trees
}
\seealso{
Other pairwise tree distances: \code{\link{AllDists}},
  \code{\link{PairwiseDistances}}
}
\author{
Martin R. Smith
}
\concept{pairwise tree distances}
