% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{distanceDistributions}
\alias{distanceDistributions}
\alias{distanceDistribution25}
\alias{distanceDistribution50}
\alias{pectinateDistances11}
\title{Distances between random pairs of trees}
\format{
Objects of class \code{matrix} (inherits from \code{array}) with
21 rows, each corresponding
to a tree distance method and is named with its abbreviation
(listed in 'Methods tested' below), and 10 000
(\code{distanceDistribution25/50})
or 100 000 (\code{pectinateDistances11}) columns,
listing the calculated distances between each pair of trees.

An object of class \code{matrix} (inherits from \code{array}) with 21 rows and 10000 columns.

An object of class \code{matrix} (inherits from \code{array}) with 21 rows and 10000 columns.

An object of class \code{matrix} (inherits from \code{array}) with 22 rows and 100000 columns.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
distanceDistribution25

distanceDistribution50

pectinateDistances11
}
\description{
\verb{distanceDistribution25(/50)} are two-dimensional matrices listing the
normalized distances between random pairs of bifurcating trees with 25 and
50 leaves drawn from the uniform distribution using
\code{ape::rtree(nTip, br = NULL)} (data object
\code{\link{randomTreePairs25}}\verb{(/50)}).
\code{pectinateDistances11} reports distances between a pectinate 11-leaf tree
and 100 000 random binary trees.
}
\section{Methods tested}{
\itemize{
\item \code{pid}: Phylogenetic Information Distance (Smith 2020), normalized
against the phylogenetic information content of the splits in the trees
being compared.
\item \code{msid}: Matching Split Information Distance (Smith 2020), normalized
against the phylogenetic information content of the splits in the trees
being compared.
\item \code{cid}: Clustering Information Distance (Smith 2020), normalized
against the entropy of the splits in the trees being compared.
\item \code{qd}: Quartet divergence (Smith 2019), normalized against its maximum
possible value for \emph{n}-leaf trees.
\item \code{nye}: Nye \emph{et al.} tree distance (Nye \emph{et al.} 2006), normalized against
the total number of splits in the trees being compared.
\item \code{jnc2}, \code{jnc4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings prohibited ('no-conflict'), normalized against
the total number of splits in the trees being compared.
\item \code{jco2}, \code{jco4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings permitted ('conflict-ok'), normalized against
the total number of splits in the trees being compared.
\item \code{ms}: Matching Split Distance (Bogdanowicz & Giaro 2012), unnormalized.
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009), unnormalized.
\item \code{masti}: Information content of Maximum Agreement Subtree, unnormalized.
\item \code{nni_l}, \code{nni_t}, \code{nni_u}: Lower bound, tight upper bound, and upper bound
for nearest-neighbour interchange distance (Li \emph{et al.} 1996), unnormalized.
\item \code{spr}: Approximate subtree prune and regraft \acronym{SPR} distance,
unnormalized.
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(\acronym{TBR}) distance, calculated using
\href{https://ms609.github.io/TBRDist/}{\pkg{TBRDist}}; unnormalized.
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1981), unnormalized.
\item \code{icrf}: Robinson-Foulds distance, splits weighted by phylogenetic
information content (Smith 2020), unnormalized.
\item \code{path}: Path distance (Steel & Penny 1993), unnormalized.
}

\itemize{
\item \code{mafi} (\code{pectinateDistances11} only): information content of the
maximum agreement forest (Smith 2020).
}
}

\references{
\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Kendall2016}{TreeDistData}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}
}
\seealso{
Tree pairs between which distances were calculated are available
in data objects \code{\link{randomTreePairs25}} and \code{\link{randomTreePairs50}}.
}
\keyword{datasets}
