% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linTests}
\alias{linTests}
\alias{linTestOneResults}
\alias{linTestSPRResults}
\alias{linTestTwoResults}
\title{Evaluating tree distance metrics with method of \emph{Lin et al.} 2012}
\format{An object of class \code{array} of dimension 5 x 16 x 5.}
\usage{
linTestOneResults

linTestSPRResults

linTestTwoResults
}
\description{
<<#TODO Describe in detail.>>
}
\references{
\insertRef{Lin2012}{TreeDistData}
}
\keyword{datasets}
