% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linTests}
\alias{linTests}
\alias{linTestOneResults}
\alias{linTestTwoResults}
\alias{linTestSPRResults}
\title{Evaluating tree distance metrics by cluster recovery}
\format{
A three-dimensional array.

Rows correspond to the clustering methods:
\itemize{
\item \code{spc}: spectral clustering
\item \code{pam}: partitioning around medioids
\item \code{h...}: hierarchical clustering using:
\code{h.cmp}, complete;
\code{h.sng}, single; and
\code{h.avg}, average linkage.
}

Columns correspond to distance metrics; see 'Methods tested' below.

Slices correspond to values of \emph{k}:
\itemize{
\item \code{linTestOneResults}: \emph{k} = 30, 40, 50, 60, 70
\item \code{linTestTwoResults}: \emph{k} = 10, 20, 30, 40
\item \code{linTestSPRResults}: \emph{k} = 30, 40, 50, 60, 70
}

An object of class \code{array} of dimension 5 x 18 x 4.

An object of class \code{array} of dimension 5 x 18 x 5.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
linTestOneResults

linTestTwoResults

linTestSPRResults
}
\description{
An effective measure of tree distance will recover clusters of similar
trees.  These datasets contain the results of tests modelled on those
in Lin \emph{et al.} (2012).
}
\details{
I used three approaches to generate clusters of similar trees, and tested
each metric in its ability to recover these clusters (Lin \emph{et al.}, 2012).

For the first test, I generated 500 datasets of 100 binary trees with
\emph{n} = 40 leaves.
Each set of trees was created by randomly selecting two \emph{k}-leaf
'skeleton' trees, where \emph{k} ranges from 0.3 \emph{n} to 0.9 \emph{n}.
From each skeleton, 50 trees were generated by adding each of the remaining
\emph{n} - \emph{k} leaves in turn at a uniformly selected point on the tree.

For the second and third test, each dataset was constructed by selecting at
random two binary 40-leaf trees.
From each starting tree, I generated 50 binary trees by conducting \emph{k}
leaf-label interchange (LLI) operations (test two) or \emph{k} subtree prune and
regraft (SPR) operations (test three) on the starting tree.
An LLI operation swaps the positions of two randomly selected leaves,
without affecting tree shape; an SPR operation moves a subtree to a new
location within the tree.

For each dataset, I calculated the distance between each pair of trees.
Trees where then partitioned into clusters using five methods,
using the \pkg{stats} and \pkg{cluster}.
I define the success rate of each distance measure as the proportion of
datasets in which every tree generated from the same skeleton was placed
in the same cluster.

For analysis of this data, see the
\href{https://ms609.github.io/TreeDistData/articles/06-lin-cluster-recovery.html}{accompanying vignette}.
}
\section{Methods tested}{
\itemize{
\item \code{dpi}: Different Phylogenetic Information (Smith, forthcoming)
\item \code{msid}: Matching Split Information Distance (Smith, forthcoming)
\item \code{cid}: Clustering Information Distance (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{ja2}, \code{ja4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings prohobited
\item \code{jna2}, \code{jna4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings permitted
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012)
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009)
\item \code{masti}: Information content of Maximum Agreement Subtree
\item \code{nni_l}, \code{nni_t}, \code{nni_u}: Lower bound, tight upper bound, and upper bound
for  nearest-neighbour interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate SPR distance
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(TBR) distance
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1985)
\item \code{rfi}: Information-corrected Robinson-Foulds distance (Smith, forthcoming)
\item \code{path}: Path distance (Steel & Penny 1993), unnormalized
}
}

\references{
\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Kendall2016}{TreeDistData}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}

\insertRef{Lin2012}{TreeDistData}
}
\keyword{datasets}
