% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PairwiseDistances}
\alias{PairwiseDistances}
\title{Distances between each pair of trees}
\usage{
PairwiseDistances(trees, Func, valueLength = 1L, ...)
}
\arguments{
\item{trees}{List of trees of class \code{phylo}.}

\item{Func}{distance function returning distance between two trees,
e.g. \link[=path.dist]{phangorn::treedist}.}
}
\value{
Matrix detailing distance between each pair of trees.
Identical trees are assumed to have zero distance.
}
\description{
Distances between each pair of trees
}
\seealso{
Other pairwise tree distances: \code{\link{AllDists}},
  \code{\link{CompareAllTrees}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{pairwise tree distances}
