% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomTreeDistances}
\alias{randomTreeDistances}
\title{Mean distances between random pairs of trees}
\format{An object of class \code{array} of dimension 20 x 13 x 197.}
\usage{
randomTreeDistances
}
\description{
A three-dimensional array listing the normalized distances between 1 000
random pairs of trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br=NULL)}.
}
\details{
Normalization is against the maximum possible value obtainable on a pair
of trees of the shapes given, with the exception of the SPR distance,
which is normalized against the upper bound (\code{spr}) and lower bound (\code{sprLB})
of its diameter on a tree with \emph{n} leaves.  The path and matching split
distances are not normalized.

Rows are named with an abbreviation of the tree comparison metric:
\itemize{
\item \code{dpi}: Variation of Phylogenetic Information (Smith, forthcoming)
\item \code{msid}: Variation of Matching Split Information (Smith, forthcoming)
\item \code{cid}: Variation of Clustering Information (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{ja2}, \code{ja4}: Jaccard-Robinson-Foulds distance (Böcker \emph{et al}. 2013),
with \emph{k} = 2, 4, with arboreal matchings enforced
\item \code{jna2}, \code{jna4}: JRF distance, non-arboreal matchings permitted
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012)
\item \code{mast}, \code{masti}: Size and information content of maximum agreement subtree
\item \code{nni_l}, \code{nni_u}: Lower and upper bounds on the nearest neighbour
interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate subtree prune and regraft distance, calculated using
\code{phangorn::SPR.dist}
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bounds on the tree bisection and
reconnection distance, calculated using
\href{https://ms609.github.io/TBRDist/}{TBRDist}
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1981)
\item \code{rfi}: Robinson-Foulds distance, splits weighted by phylogenetic
information content (Smith, forthcoming)
\item \code{path}: Path distance (Steel & Penny 1993)
}

Columns list the summary statistics of calculated tree distances: the
minimum (\code{min}),
1\%, 5\%, 10\%, 25\%, 50\% (i.e. median), 75\%, 90\%, 95\%, 99\% percentiles,
maximum (\code{max}), mean (\code{mean}) and standard deviation (\code{sd}).

The third dimension lists the number of leaves in the trees compared.
}
\references{
\itemize{
\item \insertRef{Allen2001}{TreeDist}
\item \insertRef{Bocker2013}{TreeDist}
\item \insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}
\item \insertRef{Nye2006}{TreeDist}
\item \insertRef{Robinson1981}{TreeDist}
\item \insertRef{SmithTern}{TreeSearch}
\item \insertRef{SmithDist}{TreeDist}
\item \insertRef{Steel1993}{TreeDist}
}
}
\keyword{datasets}
