% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{randomTreeDistances}
\alias{randomTreeDistances}
\title{Mean distances between random pairs of trees}
\format{An object of class \code{array} of dimension 10 x 13 x 197.}
\usage{
randomTreeDistances
}
\description{
A three-dimensional array listing the normalized distances between 1000 random
pairs of trees drawn from the uniform distribution using
\code{ape::rtree(nTip, br=NULL)}.
}
\details{
Normalization is against the maximum possible value obtainable on a pair
of trees of the shapes given, with the exception of the SPR distance,
which is normalized against the upper bound (\code{spr}) and lower bound (\code{sprLB})
of its diameter on a tree with \emph{n} tips.  The path and matching split
distances are not normalized.

Rows are named with an abbreviation of the tree comparison metric:
\itemize{
\item \code{vpi}: Variation of Phylogenetic Information (Smith, forthcoming)
\item \code{vmsi}: Variation of Matching Split Information (Smith, forthcoming)
\item \code{vci}: Variation of Clustering Information (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012)
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1981)
\item \code{path}: Path distance (Steel & Penny 1993)
\item \code{spr}: SPR distance, normalized against upper bound of diameter (Allen & Steel, 2001)
\item \code{sprLB}: SPR distance, normalized against lower bound of diameter (Allen & Steel, 2001)
}

Columns list the summary statistics of calculated tree distances: the
minimum (\code{min}),
1%, 5%, 10%, 25%, 50% (i.e. median), 75%, 90%, 95%, 99% percentiles,
maximum (\code{max}), mean (\code{mean}) and standard deviation (\code{sd}).

The third dimension lists the number of tips in the trees compared.
}
\references{
\itemize{
\item \insertRef{Allen2001}{TreeSearch}
\item \insertRef{Bogdanowicz2012}{TreeDist}
\item \insertRef{Nye2006}{TreeDist}
\item \insertRef{Robinson1981}{TreeDist}
\item \insertRef{SmithTern}{TreeSearch}
\item \insertRef{SmithDist}{TreeDist}
\item \insertRef{Steel1993}{TreeDist}
}
}
\keyword{datasets}
