% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sevenTipDistances}
\alias{sevenTipDistances}
\title{Distances between unrooted seven-leaf trees}
\format{An object of class \code{list} of length 21.}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
sevenTipDistances
}
\description{
Distances between each possible pairing of the 945 unrooted seven-leaf trees
(equivalent to rooted 6-leaf trees).  Following Kendall and Colijn (2016).
}
\details{
The list entries are named with an abbreviation of the tree comparison metric.
Information-based measures are normalized against the splitwise information
content/entropy for binary trees of the corresponding topologies.<!--#TODO VERIFY-->


The quartet distance, Nye \emph{et al.} and Jaccard-Robinson-Foulds measures
are normalized against their maximum possible values.
The remaining measures are unnormalized.
\itemize{
\item \code{dpi}: Different Phylogenetic Information (Smith, forthcoming)
\item \code{msid}: Matching Split Information Distance (Smith, forthcoming)
\item \code{cid}: Clustering Information Distance (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nts}: Nye \emph{et al.} tree similarity (Nye \emph{et al.} 2006)
\item \code{ja2}, \code{ja4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4
\item \code{jna2}, \code{jna4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
with non-arboreal matchings permitted
\item \code{msd}: Matching Split Distance (Bogdanowicz & Giaro, 2012), unnormalized
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009)
\item \code{masti}: Information content of Maximum Agreement Subtree
\item \code{nni_l}, \code{nni_t}, \code{nni_u}: Lower bound, tight upper bound, and upper bound
for  nearest-neighbour interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate SPR distance, unnormalized
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(TBR) distance
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1985), unnormalized
\item \code{rfi}: Robinson-Foulds distance, splits weighted by phylogenetic
information content (Smith, forthcoming)
\item \code{path}: Path distance (Steel & Penny 1993), unnormalized
}

Each item in the list contains a 945×945 matrix reporting the distance
between each pair of seven-leaf trees.  The first 630 trees are pectinate
(tree shape 0), the final 315 are balanced (tree shape 1).
}
\examples{
plot(TreeTools::UnrootedTreeWithShape(0, 7))
plot(TreeTools::UnrootedTreeWithShape(1, 7))

}
\references{
\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Kendall2016}{TreeDistData}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}
}
\keyword{datasets}
