% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapeEffect}
\alias{shapeEffect}
\title{Shape effect}
\format{An object of class \code{list} of length 21.}
\usage{
shapeEffect
}
\description{
...#TODO Describe in detail...
}
\keyword{datasets}
