% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{sprDistances}
\alias{sprDistances}
\title{Tree distance and SPR moves}
\format{
A list of length 21.
Each entry is named according to the corresponding tree distance method; see
'Methods tested' below.

Each member of the list is a 100 × 100 matrix listing the distance
between each pair of trees in the SPR chain (see 'Details'),
numbered from 1 to 100.
}
\source{
Scripts used to generate data objects are housed in the
\href{https://github.com/ms609/TreeDistData/tree/master/data-raw}{\code{data-raw}}
directory.
}
\usage{
sprDistances
}
\description{
Datasets testing whether separating trees by increasingly many moves
results in a corresponding increase in their distance.
}
\details{
I generated a chain of 100 50-leaf trees, starting from a pectinate tree
and deriving each tree in turn by performing an SPR operation on the previous
tree.
A consistent measure of tree similarity should correlate with the number of
SPR operations separating a pair of trees in this chain.
This said, because one SPR operation may counteract some of the difference
introduced by a previous one, perfect correlation is unlikely.

For analysis of this data, see the
\href{https://ms609.github.io/TreeDistData/articles/08-spr-walking.html}{accompanying vignette}.
}
\section{Methods tested}{
\itemize{
\item \code{pid}: Phylogenetic Information Distance (Smith, forthcoming)
\item \code{msid}: Matching Split Information Distance (Smith, forthcoming)
\item \code{cid}: Clustering Information Distance (Smith, forthcoming)
\item \code{qd}: Quartet divergence (Smith, 2019)
\item \code{nye}: Nye \emph{et al.} tree distance (Nye \emph{et al.} 2006)
\item \code{jnc2}, \code{jnc4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings prohibited ('no-conflict')
\item \code{joc2}, \code{jco4}: Jaccard-Robinson-Foulds distances with \emph{k} = 2, 4,
conflicting pairings permitted ('conflict-ok')
\item \code{ms}: Matching Split Distance (Bogdanowicz & Giaro 2012)
\item \code{mast}: Size of Maximum Agreement Subtree (Valiente 2009)
\item \code{masti}: Information content of Maximum Agreement Subtree
\item \code{nni_l}, \code{nni_t}, \code{nni_u}: Lower
bound, tight upper bound, and upper bound
for nearest-neighbour interchange distance (Li \emph{et al.} 1996)
\item \code{spr}: Approximate SPR distance
\item \code{tbr_l}, \code{tbr_u}: Lower and upper bound for tree bisection and reconnection
(TBR) distance, calculated using
\href{https://ms609.github.io/TBRDist/}{\pkg{TBRDist}}
\item \code{rf}: Robinson-Foulds distance (Robinson & Foulds 1981)
\item \code{icrf}: Information-corrected Robinson-Foulds distance (Smith, forthcoming)
\item \code{path}: Path distance (Steel & Penny 1993)
}
}

\references{
\insertRef{Bogdanowicz2012}{TreeDist}

\insertRef{Li1996}{TreeDist}

\insertRef{Kendall2016}{TreeDistData}

\insertRef{Nye2006}{TreeDist}

\insertRef{Robinson1981}{TreeDist}

\insertRef{SmithTern}{TreeSearch}

\insertRef{SmithDist}{TreeDist}

\insertRef{Steel1993}{TreeDist}

\insertRef{Valiente2009}{TreeDist}
}
\keyword{datasets}
