% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeDistData-package.R
\docType{package}
\name{TreeDistData-package}
\alias{TreeDistData}
\alias{TreeDistData-package}
\title{TreeDistData: Large Data Files to Accompany the 'TreeDist' Package}
\description{
Data used to build vignettes in the "TreeDist" package, 
  and figures in Smith (20XX).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeDistData}
  \item \url{https://github.com/ms609/TreeDistData}
  \item Report bugs at \url{https://github.com/ms609/TreeDistData/issues/}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (0000-0001-5660-1727) [copyright holder]

}
\keyword{internal}
