% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TreeDistMethods}
\alias{TreeDistMethods}
\alias{tdAbbrevs}
\alias{tdPlotSequence}
\alias{tdMdAbbrevs}
\alias{tdBoxAbbrevs}
\alias{tdMethods}
\alias{tdCol}
\alias{TDFunctions}
\alias{TDPair}
\title{Method parameters}
\format{An object of class \code{expression} of length 25.}
\usage{
tdAbbrevs

tdPlotSequence

tdMdAbbrevs

tdBoxAbbrevs

tdMethods

tdCol

TDFunctions

TDPair
}
\description{
Metadata for methods examined in this package.
}
\details{
\code{tdAbbrevs} lists abbreviations for each method, using expressions to allow
formatting of text when plotted.

\code{tdPlotSequence} lists the 20 methods discussed in the main article,
in the sequence in which they are plotted in figures.

\code{tdMdAbbrevs} uses markdown formatting.

\code{tdBoxAbbrevs} uses line breaks to fit abbreviations in an approximately
square bounding box.

\code{tdCol} provides each method with a suitable plotting colour.

\code{TDFunctions} lists for each method a function that will calculate the
distance between two trees or lists of trees.

\code{TDPair} lists for each method a function to calculate the distance
between one tree (\code{tr}) and another tree (\code{ref}).
}
\keyword{datasets}
